+++ kotest-assertions/kotest-assertions-json/src/commonMain/kotlin/io/kotest/assertions/json/compare.kt
@@ -67,6 +67,13 @@ enum class CompareOrder {
    LenientAll,
 }
 
+enum class FieldComparison {
+   /** Verifies all expected key-values match, and that no extra keys exist in the actual */
+   Exact,
+   /** Verifies all expected key-values match, ignoring keys in actual that are not specified in expected */
+   IgnoreExtra
+}
+
 /**
  * Compares two json trees, returning a detailed error message if they differ.
  */
@@ -75,15 +82,16 @@ internal fun compare(
    expected: JsonNode,
    actual: JsonNode,
    mode: CompareMode,
-   order: CompareOrder
+   order: CompareOrder,
+   fieldComparison: FieldComparison,
 ): JsonError? {
    return when (expected) {
       is JsonNode.ObjectNode -> when (actual) {
-         is JsonNode.ObjectNode -> compareObjects(path, expected, actual, mode, order)
+         is JsonNode.ObjectNode -> compareObjects(path, expected, actual, mode, order, fieldComparison)
          else -> JsonError.ExpectedObject(path, actual)
       }
       is JsonNode.ArrayNode -> when (actual) {
-         is JsonNode.ArrayNode -> compareArrays(path, expected, actual, mode, order)
+         is JsonNode.ArrayNode -> compareArrays(path, expected, actual, mode, order, fieldComparison)
          else -> JsonError.ExpectedArray(path, actual)
       }
       is JsonNode.BooleanNode -> compareBoolean(path, expected, actual, mode)
@@ -99,19 +107,22 @@ internal fun compareObjects(
    actual: JsonNode.ObjectNode,
    mode: CompareMode,
    order: CompareOrder,
+   fieldComparison: FieldComparison,
 ): JsonError? {
 
-   val keys1 = expected.elements.keys
-   val keys2 = actual.elements.keys
+   if (fieldComparison == FieldComparison.Exact) {
+      val keys1 = expected.elements.keys
+      val keys2 = actual.elements.keys
 
-   if (keys1.size < keys2.size) {
-      val missing = keys2 - keys1
-      return JsonError.ObjectMissingKeys(path, missing)
-   }
+      if (keys1.size < keys2.size) {
+         val missing = keys2 - keys1
+         return JsonError.ObjectMissingKeys(path, missing)
+      }
 
-   if (keys2.size < keys1.size) {
-      val extra = keys1 - keys2
-      return JsonError.ObjectExtraKeys(path, extra)
+      if (keys2.size < keys1.size) {
+         val extra = keys1 - keys2
+         return JsonError.ObjectExtraKeys(path, extra)
+      }
    }
 
    // when using strict order mode, the order of elements in json matters, normally, we don't care
@@ -119,7 +130,7 @@ internal fun compareObjects(
       CompareOrder.Strict ->
          expected.elements.entries.withIndex().zip(actual.elements.entries).forEach { (e, a) ->
             if (a.key != e.value.key) return JsonError.NameOrderDiff(path, e.index, e.value.key, a.key)
-            val error = compare(path + a.key, e.value.value, a.value, mode, order)
+            val error = compare(path + a.key, e.value.value, a.value, mode, order, fieldComparison)
             if (error != null) return error
          }
       CompareOrder.LenientAll,
@@ -127,7 +138,7 @@ internal fun compareObjects(
       CompareOrder.Lenient ->
          expected.elements.entries.forEach { (name, e) ->
             val a = actual.elements[name] ?: return JsonError.ObjectMissingKeys(path, setOf(name))
-            val error = compare(path + name, e, a, mode, order)
+            val error = compare(path + name, e, a, mode, order, fieldComparison)
             if (error != null) return error
          }
    }
@@ -141,6 +152,7 @@ internal fun compareArrays(
    actual: JsonNode.ArrayNode,
    mode: CompareMode,
    order: CompareOrder,
+   fieldComparison: FieldComparison,
 ): JsonError? {
 
    if (expected.elements.size != actual.elements.size)
@@ -151,7 +163,7 @@ internal fun compareArrays(
       CompareOrder.Lenient,
       CompareOrder.Strict -> {
          expected.elements.withIndex().zip(actual.elements.withIndex()).forEach { (a, b) ->
-            val error = compare(path + "[${a.index}]", a.value, b.value, mode, order)
+            val error = compare(path + "[${a.index}]", a.value, b.value, mode, order, fieldComparison)
             if (error != null) return error
          }
       }
@@ -172,7 +184,7 @@ internal fun compareArrays(
          fun findMatchingIndex(element: JsonNode): Int? {
             for (i in availableIndexes()) {
                // Comparison with no error -> matching element
-               val isMatch = compare(path + "[$i]", element, expected.elements[i], mode, order) == null
+               val isMatch = compare(path + "[$i]", element, expected.elements[i], mode, order, fieldComparison) == null
 
                if (isMatch) {
                   return i